migrate gql query to /gql dir
